Support Whitespace tags for Objects
This patch is the first step towards supporting the Whitespace Control tags.
It only implements support for the whitespace object tags `{{-` and `-}}`.

Support for the tags is more complicated since the `Tag.eval/3` function needs to support the `trim_previous` and `trim_next` arguments.
In some cases there are multiple of them and we need to forward them to further calls of `Solid.render/3`.

I'm very new to NimbleParsec and generally parsers so I'm posting this to get verification on the current path I have taken.
Some guidance would be appreciated if possible.

Needed by https://github.com/edgurgel/solid/issues/10