Tests, Add New assert_render helper
While implementing [whitespace support](https://github.com/edgurgel/solid/pull/67) I came to the conclusion that it
makes sense to auto generate different whitespace trimming variants of
a given template to gain a good test coverage of edge cases while keeping test maintainable.

One problem I encountered, was that since the templates where generated
I didn't know the actual input that was used when comparing the render results of liquid and solid.

My solution was adding a new `assert_render` macro that renders the template in liquid and solid
and then compares them (as we do already with `cases_test`), but then includes the passed template in the assertion output if there is an missmatch.

I'm trying to keep the main patch for the whitespace support as small and revieable as possible.
Since this particular change can stand alone and IMHO makes also sense outside of the context of whitespace support I've opened this PR.

Sample output:
```
  1) test case 113 (:"Elixir.Solid.Integration.Cases.113Test")
     test/integration/cases_test.exs:17
     Render result was different!
     Input:
     {% if enabled? %}
     ON
     {% else %}
     OFF
     {% endif %}

     code:  liquid_output == solid_output
     left:  "\nON\n\n"
     right: "\nOFF\n\n"
     stacktrace:
       test/integration/cases_test.exs:21: (test)
```