Allow custom filter module to be passed in through render option
When using this module in different parts of a codebase, one might
encounter the situation that a certain template needs a different set
of filters available than other templates.

The current approach of using application environment only allows for
a single custom filter module to be defined globally. This patch
changes that by creating a new option, `custom_filters:`, to be passed
in at render time.